web/requirements.txt
Flask
Flask-Session
gunicorn
python-dotenv
redis
bcrypt
pyjwt
requests